requirements.txt
Flask
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
mysql-connector
uwsgi